+++ addons/african_militia/CfgVehicles_Turrets_O.hpp
@@ -22,11 +22,3 @@ class TACU_African_Militia_V_O_ZU23: TACU_Assets_ZU23_Base {
     typicalCargo[] = {"TACU_African_Militia_U_O_Rifleman"};
     textureList[] = {"SFIA", 1};
 };
-
-class TACU_African_Militia_V_O_DSHKM: TACU_Assets_DSHKM_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_African_Militia_O";
-    side = 0;
-    crew = "TACU_African_Militia_U_O_Rifleman";
-    typicalCargo[] = {"TACU_African_Militia_U_O_Rifleman"};
-};
