+++ addons/african_militia/subconfig_cup/CfgVehicles.hpp
@@ -1,0 +1,17 @@
+class CfgVehicles {
+    #include "\x\tacu\addons\assets\script_classes_vehicles_cup.hpp"
+
+    class TACU_African_Militia_V_O_DSHKM: TACU_Assets_DSHKM_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_African_Militia_O";
+        side = 0;
+        crew = "TACU_African_Militia_U_O_Rifleman";
+        typicalCargo[] = {"TACU_African_Militia_U_O_Rifleman"};
+    };
+    class TACU_African_Militia_V_I_DSHKM: TACU_African_Militia_V_O_DSHKM {
+        faction = "TACU_African_Militia_I";
+        side = 2;
+        crew = "TACU_African_Militia_U_I_Rifleman";
+        typicalCargo[] = {"TACU_African_Militia_U_I_Rifleman"};
+    };
+};
